New webscraping exercise
The Ryman site has updated to a new layout, which is significantly more
difficult to scrape.
This contains a basic webscraping exercise with a page which is not
likely to be changed.
For students who complete the basic exercise, there is a bonus site to
scrape and a challenge site.